requirements.txt
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2
diffusers==0.33.0
transformers>=4.43.0
accelerate>=0.33.0
safetensors>=0.4.3
pillow>=10.3.0
imageio[ffmpeg]>=2.34.0
opencv-python-headless>=4.10.0.84
numpy>=1.26.4,<2
streamlit>=1.36.0
fastapi>=0.110
uvicorn[standard]>=0.30.0
requests>=2.32.0

